feat 3.10: fishing adventures CRUD and search, instructors availability and profile editing
Plus some small changes and adjustments

+++ BookingApplication/src/main/java/com/BookingApp/dto/PricelistItemDto.java
@@ -4,9 +4,9 @@ public class PricelistItemDto {
 
 	public long instructorsId;
 	public String description;
-	public long price;
+	public double price;
 	
-	public PricelistItemDto(long instructorsId, String description, long price) {
+	public PricelistItemDto(long instructorsId, String description, double price) {
 		super();
 		this.instructorsId = instructorsId;
 		this.description = description;
